UPD: FE - Refactor gameboard -> playground
